refactor: split apart component mutation observer logic into standalone JS utility in Bolt core
