+++ packages/core/utils/index.js
@@ -1,4 +1,5 @@
 export * from './color-contrast';
+export * from './contains-any';
 export * from './css';
 export * from './declarative-click-handler';
 export * from './environment';
@@ -10,6 +11,7 @@ export * from './supports-css-vars';
 export * from './supports-passive-event-listener';
 export * from './which-transition-event';
 export * from './wait-for-transition-end';
+export * from './watch-for-component-mutations';
 
 // https://www.polymer-project.org/3.0/docs/devguide/custom-elements#defer-work
 export { afterNextRender } from '@polymer/polymer/lib/utils/render-status.js';
